refactor: Remove unnecessary code and ensuring only code being used is included
Removed a lot of factory code that was not getting used to reduce
clutter
