--- src/js/dom/KL.DomUtil.js
@@ -1,157 +1,0 @@
-/*	KL.DomUtil
-    Inspired by Leaflet
-    KL.DomUtil contains various utility functions for working with DOM
-================================================== */
-
-
-module.exports = {
-    get: function (id) {
-        return (typeof id === 'string' ? document.getElementById(id) : id);
-    },
-
-    getStyle: function (el, style) {
-        var value = el.style[style];
-        if (!value && el.currentStyle) {
-            value = el.currentStyle[style];
-        }
-        if (!value || value === 'auto') {
-            var css = document.defaultView.getComputedStyle(el, null);
-            value = css ? css[style] : null;
-        }
- 
-        return (value === 'auto' ? null : value);
-    },
-
-    getViewportOffset: function (element) {
-        var top = 0,
-        left = 0,
-        el = element,
-        docBody = document.body;
-
-        do {
-            top += el.offsetTop || 0;
-            left += el.offsetLeft || 0;
-
-            if (el.offsetParent === docBody &&
-                KL.DomUtil.getStyle(el, 'position') === 'absolute') {
-                break;
-            }
-            el = el.offsetParent;
-        } while (el);
-
-        el = element;
-
-        do {
-            if (el === docBody) {
-                break;
-            }
-
-            top -= el.scrollTop || 0;
-            left -= el.scrollLeft || 0;
-
-            el = el.parentNode;
-        } while (el);
-
-        return new KL.Point(left, top);
-    },
-
-    create: function (tagName, className, container) {
-        var el = document.createElement(tagName);
-        el.className = className;
-        if (container) {
-            container.appendChild(el);
-        }
-
-        return el;
-    },
-
-    disableTextSelection: function () {
-        if (document.selection && document.selection.empty) {
-            document.selection.empty();
-        }
-        if (!this._onselectstart) {
-            this._onselectstart = document.onselectstart;
-            document.onselectstart = KL.Util.falseFn;
-        }
-    },
-
-    enableTextSelection: function () {
-        document.onselectstart = this._onselectstart;
-        this._onselectstart = null;
-    },
-
-    hasClass: function (el, name) {
-        return (el.className.length > 0) &&
-            new RegExp("(^|\\s)" + name + "(\\s|$)").test(el.className);
-    },
-
-    addClass: function (el, name) {
-        if (!KL.DomUtil.hasClass(el, name)) {
-            el.className += (el.className ? ' ' : '') + name;
-        }
-    },
-
-    removeClass: function (el, name) {
-        el.className = el.className.replace(/(\S+)\s*/g, function (w, match) {
-            if (match === name) {
-                return '';
-            }
-        
-            return w;
-        }).replace(/^\s+/, '');
-    },
-
-    setOpacity: function (el, value) {
-        if (KL.Browser.ie) {
-            el.style.filter = 'alpha(opacity=' + Math.round(value * 100) + ')';
-        } else {
-            el.style.opacity = value;
-        }
-    },
-
-    testProp: function (props) {
-        var style = document.documentElement.style;
-
-        for (var i = 0; i < props.length; i++) {
-            if (props[i] in style) {
-                return props[i];
-            }
-        }
-        
-        return false;
-    },
-
-    getTranslateString: function (point) {
-
-        return KL.DomUtil.TRANSLATE_OPEN +
-            point.x + 'px,' + point.y + 'px' +
-            KL.DomUtil.TRANSLATE_CLOSE;
-    },
-
-    getScaleString: function (scale, origin) {
-        var preTranslateStr = KL.DomUtil.getTranslateString(origin),
-            scaleStr = ' scale(' + scale + ') ',
-            postTranslateStr = KL.DomUtil.getTranslateString(origin.multiplyBy(-1));
-
-        return preTranslateStr + scaleStr + postTranslateStr;
-    },
-
-    setPosition: function (el, point) {
-        el._vco_pos = point;
-        if (KL.Browser.webkit3d) {
-            el.style[KL.DomUtil.TRANSFORM] =  KL.DomUtil.getTranslateString(point);
-
-            if (KL.Browser.android) {
-                el.style['-webkit-perspective'] = '1000';
-                el.style['-webkit-backface-visibility'] = 'hidden';
-            }
-        } else {
-            el.style.left = point.x + 'px';
-            el.style.top = point.y + 'px';
-        }
-    },
-
-    getPosition: function (el) {
-        return el._vco_pos;
-    }
-};
