+++ src/js/quote/KL.QuoteComposition.js
@@ -1,10 +1,21 @@
 /*	KL.QuoteComposition
 ================================================== */
 
+KL.Class = require("core/KL.Class");
+
 module.exports = KL.Class.extend({
 
     includes: [KL.Events, KL.DomMixins],
     _el: {},
+
+    create: function(tagName, className, container) {
+        var el = document.createElement(tagName);
+        el.className = className;
+        if (container) {
+            container.appendChild(el);
+        }
+        return el;
+    },
 
     /*	Constructor
     ================================================== */
@@ -50,7 +61,7 @@ module.exports = KL.Class.extend({
         KL.Util.mergeData(this.options, options);
         KL.Util.mergeData(this.data, data);
 
-        this._el.container = KL.Dom.create("div", this.options.base_classname);
+        this._el.container = create("div", this.options.base_classname);
 
         this._updateClassName();
 
@@ -70,7 +81,7 @@ module.exports = KL.Class.extend({
     /*	Events
     ================================================== */
     _onMouseClick: function() {
-        this.fireEvent("clicked", this.options);
+        return this;
     },
 
     _onContentEdit: function() {
@@ -143,10 +154,6 @@ module.exports = KL.Class.extend({
         // });
     },
 
-    _onLoaded: function() {
-        this.fireEvent("loaded", this.options);
-    },
-
     /*	Private Methods
     ================================================== */
     _determineTextSize: function(q) {
@@ -215,17 +222,17 @@ module.exports = KL.Class.extend({
     _initLayout: function () {
 
         // Create Layout
-        this._el.composition_container 	= KL.Dom.create("div", "kl-quotecomposition-container", this._el.container);
-        this._el.composition_text = KL.Dom.create("div", "kl-quotecomposition-text", this._el.composition_container);
-        this._el.blockquote	= KL.Dom.create("blockquote", "", this._el.composition_text);
-        this._el.blockquote_p = KL.Dom.create("p", "", this._el.blockquote);
-        this._el.citation = KL.Dom.create("cite", "", this._el.blockquote);
-        this._el.background	= KL.Dom.create("div", "kl-quotecomposition-background", this._el.composition_container);
-        this._el.image = KL.Dom.create("div", "kl-quotecomposition-image", this._el.composition_container);
+        this._el.composition_container 	= create("div", "kl-quotecomposition-container", this._el.container);
+        this._el.composition_text = create("div", "kl-quotecomposition-text", this._el.composition_container);
+        this._el.blockquote	= create("blockquote", "", this._el.composition_text);
+        this._el.blockquote_p = create("p", "", this._el.blockquote);
+        this._el.citation = create("cite", "", this._el.blockquote);
+        this._el.background	= create("div", "kl-quotecomposition-background", this._el.composition_container);
+        this._el.image = create("div", "kl-quotecomposition-image", this._el.composition_container);
 
         // Create Buttons
-        this._el.button_group = KL.Dom.create("div", "kl-button-group", this._el.container);
-        this._el.button_download = KL.Dom.create("a", "kl-button kl-button-right", this._el.button_group);
+        this._el.button_group = create("div", "kl-button-group", this._el.container);
+        this._el.button_download = create("a", "kl-button kl-button-right", this._el.button_group);
 
         this._el.button_download.innerHTML = "Save";
